Maintenance/146/ongoing code maintenance

Closes #146

See merge request azriel91/autexousious!147

+++ crate/game_input/src/component/controller_input.rs
@@ -10,13 +10,13 @@ pub struct ControllerInput {
     /// * 0.0 if neither or both left and right buttons are pressed.
     /// * -1.0 if only left button is pressed.
     /// * 1.0 if only right button is pressed.
-    pub x_axis_value: f64,
+    pub x_axis_value: f32,
     /// Z axis input value.
     ///
     /// * 0.0 if neither or both up and down buttons are pressed.
     /// * -1.0 if only up button is pressed.
     /// * 1.0 if only down button is pressed.
-    pub z_axis_value: f64,
+    pub z_axis_value: f32,
     /// Whether the `Defend` button is pressed.
     pub defend: bool,
     /// Whether the `Jump` button is pressed.
